[One .NET] simplify native library inclusion
I've noticed the `_IncludeNativeSystemLibraries` MSBuild target can
take 40ms because it is running this workaround:

    <!-- HACK remove any unexpected runtime components -->
    <FrameworkNativeLibrary
        Remove="@(ResolvedFileToPublish)"
        Condition="  $([System.Text.RegularExpressions.Regex]::IsMatch('%(FullPath)', '.*linux-.*'))  "
    />

The `_IncludeNativeSystemLibraries` also can be split up between
"legacy" Xamarin.Android and .NET 6. The .NET 6 version can simply
check the `@(ResolvedFileToPublish)` item group, as the other parts
support the intepreter, profiler, etc., which are not available in .NET
6 yet.

I created a new `<ProcessNativeLibraries/>` MSBuild task to simplify
the work in C#:

* We can iterate over all `.so` files from `@(ResolvedFileToPublish)`.
* We can fix up `libmonodroid.so` as needed.
* We can emit a warning if the ABI is unknown without using a `Regex`.
  It is possible for Linux `.so` files to be brought in from NuGet
  packages as discussed in: https://github.com/dotnet/runtime/pull/45927

The warning is a new alternate message for XA4301, such as:

    warning XA4301: Cannot determine ABI of native library 'C:\src\xamarin-android\packages\sqlitepclraw.lib.e_sqlite3.linux\1.1.11\runtimes\linux-arm64\native\libe_sqlite3.so'. Remove the 'SQLitePCLRaw.lib.e_sqlite3.linux' NuGet package, or notify the library author.

As part of this change, .NET 6 will no longer need to do:

    <PropertyGroup>
      <_Assemblies>@(_ResolvedFrameworkAssemblies)</_Assemblies>

So we can avoid creating a ~14K character string.

~~ Results ~~

Timing the project fromm the `DotNetBuild(android.21-arm,false)` test,
in a build with no changes:

    Before:
    40 ms  _IncludeNativeSystemLibraries              1 calls
    After:
     4 ms  _IncludeNativeSystemLibraries              1 calls

This should save ~36ms from all .NET 6 builds.