[One .NET] cleanup Microsoft.Android.Sdk.BuildOrder.targets (#5382)
Context: https://github.com/dotnet/msbuild/pull/5499

In .NET 6, we can rely on `$(IncrementalCleanDependsOn)`. This means
we can remove the expression:

    <CoreBuildDependsOn>
      $([MSBuild]::Unescape($(CoreBuildDependsOn.Replace('IncrementalClean;', '$(_BeforeIncrementalClean);IncrementalClean;'))))
    </CoreBuildDependsOn>

And simply use `$(IncrementalCleanDependsOn)` instead of
`$(_BeforeIncrementalClean)`.

I also found a block of targets defined with:

    <!-- Targets that run unless we are running _ComputeFilesToPublishForRuntimeIdentifiers -->
    <PropertyGroup Condition=" '$(_ComputeFilesToPublishForRuntimeIdentifiers)' != 'true' ">

But this file is no longer imported at all:

    <Import Project="Microsoft.Android.Sdk.BuildOrder.targets" Condition=" '$(_ComputeFilesToPublishForRuntimeIdentifiers)' != 'true' " />

So we can delete the `Condition` now.