[build] fix Inputs/Outputs for monodroid.csproj (#5426)
On macOS, I noticed this happening on every build:

	$ msbuild src/monodroid -clp:performancesummary
	...
	 2934 ms  _BuildHostRuntimes
	41768 ms  _BuildAndroidRuntimes

The original cause was:

	Building target "_BuildAndroidRuntimes" completely.
	Output file "obj/Debug//armeabi-v7a-Debug/libxamarin-app.so" does not exist.

I updated the `Outputs` to use the correct file path:

	-@(AndroidSupportedTargetJitAbi->'$(IntermediateOutputPath)\%(Identity)-Debug\libxamarin-app.so');
	-@(AndroidSupportedTargetJitAbi->'$(IntermediateOutputPath)\%(Identity)-Release\libxamarin-app.so')
	+@(AndroidSupportedTargetJitAbi->'$(OutputPath)\%(Identity)\Debug\libxamarin-app.so');
	+@(AndroidSupportedTargetJitAbi->'$(OutputPath)\%(Identity)\Release\libxamarin-app.so')

Next, I was still hitting the problem, but instead hit:

	Building target "_BuildAndroidRuntimes" completely.
	Input file "jni/*.cc" does not exist.

You can't actually use wildcards in `Inputs`? I made a new
`_FindMonoDroidSources` MSBuild target that created a new
`@(_MonoDroidSources)` item group. I used this in place of wildcards.

The `_BuildHostRuntimes` MSBuild target had a typo which caused:

	Building target "_BuildHostRuntimes" completely.
	Input file "host-Darwin-Debug/CMakeCache.txt" does not exist.

The `Inputs` needed to change:

	-$(IntermediatePath)
	+$(IntermediateOutputPath)

It also needed to use `@(_MonoDroidSources)` as `Inputs`.

Now things seem to work correctly:

	Skipping target "_BuildAndroidRuntimes" because all outputs are up-to-date with respect to the input files.
	Skipping target "_BuildHostRuntimes" because all outputs are up-to-date with respect to the input files.

This saves ~45s off of incremental builds for me on macOS.
The savings on Windows would not be as high, as it does not run the
"checked" builds at all.