[One .NET] proposal for "Single Project" (#5397)
One theme for Xamarin in .NET 6 is to simplify the cross-platform
experience between iOS and Android. We would like to provide
cross-platform alternatives to `AndroidManifest.xml` and `Info.plist`.
The eventual goal would be to remove the need for these files for
simple apps and project templates. Developers would only need to
create them when a less widely used feature is needed. We would need
to create item templates for `AndroidManifest.xml` and `Info.plist`
when (and if) they are removed from project templates.

The concepts for a Xamarin "single project" go much beyond this. This
document is a starting point for what needs to go in the .NET Android
and Apple SDKs, while much of the remaining features would be in the
Xamarin.Forms/MAUI MSBuild targets.

This is a spec for the first phase that would introduce a few MSBuild
properties for the application title, id, and version. A
cross-platform `Directory.Build.props` could be:

	<Project>
	  <PropertyGroup>
	    <ApplicationTitle>Hello!</ApplicationTitle>
	    <ApplicationId>com.companyname.hello</ApplicationId>
	    <ApplicationVersion>1.0.0</ApplicationVersion>
	  </PropertyGroup>
	</Project>

When this is implemented, we decide if we want to move ahead with the
remaining MSBuild properties that would be needed to completely omit
`Info.plist` and `AndroidManifest.xml`.