[Xamarin.Android.Build.Tasks] Fix up ZipArchiveEx to handle updates. (#5266)
While investigating another bug, I found that the code to look up an
existing entry within a zip file was not working at all.

The problem was the filename we were looking up included a `/` at the
start of the path.  The zip entires in the zip do not include a `/`
at the start, so the code was never finding a match.  As a result we
were always trying to add a file rather than skipping those which
were up to date.

In addition I found that we do not always need to create entries in
the zip for directories.  When we create the `classes.zip` in the
`<Javac/>` task we don't need to create directory entries.

Additionally, we should have the ability to turn the auto flush on/off.

There seems to be a small performance boost with these changes.

  * Old:

        749 ms  Javac
        746 ms  Javac

  * New:

        728 ms  Javac
        724 ms  Javac