[One .NET] put bundletool.jar in separate workload pack
Context: https://github.com/xamarin/xamarin-android/issues/5438

This only implements half of #5438.

`bundletool.jar` is quite large:

    24755974 bundletool.jar

Since this isn't needed for every build, only when
`$(AndroidPackageFormat)` is `aab`, we can put `bundletool.jar` in its
own "workload" pack.

The way this works is for `WorkloadManifest.targets` to define an
optional Sdk import:

    <Import Project="Sdk.targets" Sdk="Microsoft.Android.Sdk.BundleTool"
        Condition=" '$(AndroidPackageFormat)' == 'aab' " />

If the SDK is missing, for a build using `AndroidPackageFormat=aab`
you get the error:

    Microsoft.NET.Workload.Android\WorkloadManifest.targets(4,33): error MSB4236: The SDK 'Microsoft.Android.Sdk.BundleTool' specified could not be found.

Our .NET 6 installers will include `Microsoft.Android.Sdk.BundleTool`
and install it to the correct location to avoid this error.

Down the road, `Microsoft.Android.Sdk.BundleTool` would be acquired by
a future `dotnet workload` command that does not exist yet.

The resulting size of the two Sdk packs are:

    22881359 Microsoft.Android.Sdk.BundleTool.11.0.200.nupkg
    49970851 Microsoft.Android.Sdk.win-x64.11.0.200.nupkg