[xaprepare] Use global NuGet package cache instead of in-tree one
Our `NuGet.config` specifies a local path to store the nuget packages
because some of our projects need to access various data from the
packages and they need to know where to find the unpacked content.
However, the `packages/` directory being in the `Xamarin.Android`
directory means that every time one cleans the repository to build from
scratch, all the packages are removed and the next build has to
redownload them.

This commit removes the need to use local package cache, instead
allowing us to use the global cache.  Cache location is no longer
specified in `NuGet.config` but, instead, it defaults to the global
location (useful when running certain workloads on CI where `xaprepare`
is not ran) while `xaprepare` detects its location and generates a
file (`bin/Build$(Configuration)/Configuration.Generated.props`) which
defines the `XAPackagesDir` MSBuild variable first to the value of the
`NUGET_PACKAGES` environment variable ([see][0]) and if it isn't present,
to the value detected from the location of the `LibZipSharp` package on
the filesystem.  `Configuration.props`, in turn, includes the generated
file and, should `$(XAPackagesDir)` be empty, it defaults its value to
the OS-specific location.

Furthermore, the commit adds two wrapper scripts to run the NUnit
console runner: `build-tools/scripts/nunit3-console` (for Unix) and
`build-tools/scripts/nunit3-console.cmd` (for Windows).  Both scripts
first look at the `NUGET_PACKAGES` environment variable, then check if
the copy of `NuGet.exe` downloaded by `xaprepare` is available and, if
so, use it to detect the location.  Failing that, they use the default
OS-specific location of the package cache.  This removes the need to
hardcode package version into our azure pipeline as well as VSCode
configuration.  However, both scripts still have the
`nunit.consolerunner` package version embedded in them.  This is done so
that we don't need to generate them in scenarios where `xaprepare` is
undesirable to run.

[0]: https://docs.microsoft.com/en-us/nuget/consume-packages/managing-the-global-packages-and-cache-folders