[Xamarin.Android.Build.Tasks] fix @(AndroidLibrary) in legacy Xamarin.Android (#5528)
I was testing the new `@(AndroidLibrary)` item group, and noticed some
logic was backwards:

    <EmbeddedJar          Include="@(AndroidLibrary)" Condition=" '%(AndroidLibrary.Extension)' == '.jar' and '%(AndroidLibrary.Bind)' != 'true' " />
    <EmbeddedReferenceJar Include="@(AndroidLibrary)" Condition=" '%(AndroidLibrary.Extension)' == '.jar' and '%(AndroidLibrary.Bind)' == 'true' " />

`@(EmbeddedJar)` should be used when `%(Bind)` is `true`! Because it
is using `@(EmbeddedReferenceJar)` instead, we are not currently
generating C# bindings by default.

We had a test for this item group, but it didn't actually check the
output assembly has a C# class from the `@(AndroidLibrary)`.

I updated the test and fixed the logic here.