[Xamarin.Android.Build.Tasks] run 'mono generator.exe' in some cases (#5522)
Context: https://github.com/xamarin/Xamarin.Legacy.Sdk

When building this project under `dotnet build`:

	<Project Sdk="Xamarin.Legacy.Sdk">
	  <PropertyGroup>
	    <TargetFrameworks>monoandroid11.0;net6.0-android</TargetFrameworks>
	  </PropertyGroup>
	</Project>

We are in a situation where `AndroidDotnetToolTask` will attempt to
run `generator.exe` for a "legacy" Xamarin.Android binding project.

This works on Windows, because you can just run `generator.exe`.

On macOS, however, we need to run `mono generator.exe`.  If the
runtime is Mono, this happens implicitly:

  * https://github.com/mono/mono/blob/2020-02/mono/metadata/w32process-unix.c#L1745
  * https://github.com/mono/mono/blob/2020-02/mono/metadata/w32process-unix.c#L1236-L1363

But if the runtime is .NET 6, you get the error:

	/Library/Frameworks/Xamarin.Android.framework/Libraries/xbuild/Xamarin/Android/Xamarin.Android.Bindings.ClassParse.targets(35,5):
	error MSB6003: The specified task executable "generator.exe" could not be run. System.ComponentModel.Win32Exception (13): Permission denied
	  at System.Diagnostics.Process.ForkAndExecProcess(String filename, String[] argv, String[] envp, String cwd, Boolean redirectStdin, Boolean redirectStdout, Boolean redirectStderr, Boolean setCredentials, UInt32 userId, UInt32 groupId, UInt32[] groups, Int32& stdinFd, Int32& stdoutFd, Int32& stderrFd, Boolean usesTerminal, Boolean throwOnNoExec)
	  at System.Diagnostics.Process.StartCore(ProcessStartInfo startInfo)
	  at System.Diagnostics.Process.Start()
	  at Microsoft.Build.Utilities.ToolTask.ExecuteTool(String pathToTool, String responseFileCommands, String commandLineCommands)
	  at Microsoft.Build.Utilities.ToolTask.Execute() [/Users/runner/work/1/s/samples/JavaBinding/JavaBinding.csproj]

`generator.exe` won't have the execute bit set (`chmod`).  We need to
detect this situation and run `mono generator.exe` instead.

When:

 1. `generator.dll` is not found
 2. The OS is not Windows
 3. `RuntimeInformation.FrameworkDescription` does not start with `Mono`

Look in `$PATH` for a path to `mono` and use it.

If no `mono` is found in `$PATH`, try one of the "fallback" paths:

  * `/Library/Frameworks/Mono.framework/Versions/Current/Commands/mono`
  * `/usr/bin/mono`

If neither of those exist, use `mono`.  This will likely fail, but
will fail "later", when `mono generator.exe` is executed.