[tools] Add a tool to decompress XA-compressed assemblies
The utility is named `decompress-assemblies` and takes as its
parameters a list of files to decompress.  The files might be either
`.apk` archives or `.dll` files.

I case of `.apk`, managed assemblies stored in the `assemblies/`
directory inside the APK will be read and decompressed.

The uncompressed result is stored in the current directory with the
original assembly name modified to contain the `uncompressed-` prefix,
in order to prevent accidental overwriting of assemblies found in the
current directory.