[MSBuildDeviceIntegration] Give each test a unique package name.
One of the problems we might see on CI is when two tests run at the same
time on the same emulator. This can cause collisions and tests to fail
because the app was changed while the test was running.
Currently ALL of the tests user `UnnamedProject.UnnamedProject` as a
package name. This means they overwrite each other.

The commit introduces a new method `GeneratePackageName` which will
auto generate a package name based on

`com.xamarin.test<Test ID>`

where `<Test ID>` is the value we get from `TestContext.CurrentContext.Test.ID`.
We have to switch out the hyphens in the Test ID to get a valid package
name, but it will be unique. The result will be each test will have its
own package and install location. None of the tests will overwrite the
others. Addition code has been put in place for test `TearDown` to
ensure that the package is uninstalled. This is so we don't fill up the
device and run out of disk space.

From now on Device based tests should use the new method to set the
`PackageName` property.