[illink] fix for .NET 6 & AndroidLinkMode=Full
Context: https://github.com/xamarin/xamarin-macios/pull/10552

After bumping .NET 6 in 240a3e1c, I found the
`CustomLinkDescriptionPreserve(Full)` test fails:

    Linker test app did not run successfully.
    Expected: True
    But was:  False

This fails because we are producing an `.apk` file where the
`MainActivity` is completely linked away.

It seems that removing `@(TrimmerRootAssembly)` completely in 240a3e1c
only worked if `AndroidLinkMode` was `SdkOnly` and not `Full`.

I brought over the change that worked for Xamarin.iOS to fix this
issue.